Payload support
Make it so a worker can recognize a payload that's plugged in and send
a GCS message so the app knows it's been added.

Make it so you can interrogate a payload with GET /payload/check.

Make it so it detects when a payload has been unplugged, and notifies
the GCS.