Enable sending mavlink from workers
Since the rewrite, there was never an occasion to send
mavlink from a worker. I discovered that the process of
sending a mavlink message across a process boundary
doesn't actually work, since the functions on the sent object
are stripped off.

So this fix looks up the relevant mavlink message constructor
by name based on the incoming message's name, builds an argument
list, constructs a new *real* message based on it, and sends that.