Avoid starting disabled workers
Disabled workers are still loaded and their
onLoad() functions are called, but they're expected
not to spin up child processes, etc if they're
not enabled (as reported by ATTRS.enabled in onLoad()).
Instead, they should implement
onEnabledChanged(enabled) and start or stop
child processes/etc based on whether they're
becoming enabled or disabled. onEnabledChanged()
will be called when the worker is loaded, and
when its enabled state is changed from the web UI
or endpoint call.